@@ -12,11 +12,10 @@
   "dependencies": {
     "bootstrap": "^4.3.1",
     "js-cookie": "^2.2.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-redux": "^7.1.0",
-    "reactstrap": "^8.0.0",
-    "redux": "^4.0.1",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-redux": "^7.2.0",
+    "redux": "^4.0.5",
     "redux-saga": "^1.0.4"
   },
   "devDependencies": {
